refactor(tests): 공통 픽스처를 conftest.py로 통합
중복 제거(TIER1-2):
- _reset_sse_global(autouse) 7개 파일 복붙 → conftest 1개.
- client(StaticPool 인메모리 DB) 4개 파일 복붙 → conftest 1개.
  test_admin/api/api_tabs의 DB 없는 단순 client는 로컬 오버라이드로 유지.
- auth 헬퍼는 파일마다 이름·이메일·시그니처가 달라 공통화 안 함.
- os.environ["API_SKIP_INIT"] 설정은 import 타이밍 때문에 각 파일 유지.

약 100줄 중복 제거. 전체 895 테스트 green.
(test_push는 본문이 db.SessionLocal()을 직접 써서 import api.db 유지 필요 —
테스트가 즉시 잡아냄.)
